Refine node.(#42)
