+++ tests/raft_node_test.cpp
@@ -4,7 +4,7 @@
 #define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
 
 #include "doctest.h"
-#include "../src/raft_node.hpp"
+#include "../src/node/raft_node.h"
 
 //nanolog INFO (WARN) and doctest INFO (WARN) conflict
 #ifdef INFO
